Extract all layout classes to separate, acyclic modules
